Beams deal reduced damage to enemies

+++ enemy.go
@@ -14,10 +14,17 @@ type Enemy struct {
 	Sprite    *ebiten.Image
 	Alive     bool
 	Collision Vector4
-	Health    int
+	Health    float64
 }
 
-func NewEnemy(screenWidth, screenHeight int, speed float64, sprite *ebiten.Image, player *Player, alive bool, health int) *Enemy {
+func NewEnemy(
+	screenWidth, screenHeight int,
+	speed float64,
+	sprite *ebiten.Image,
+	player *Player,
+	alive bool,
+	health float64,
+) *Enemy {
 	enemyBounds := sprite.Bounds()
 
 	enemyWidth := float64(enemyBounds.Dx())
@@ -86,7 +93,6 @@ func (e *Enemy) CollisionBounds() Vector4 {
 }
 
 func (e *Enemy) Update(p *Player, blocks []*Block) {
-
 	if e.Health <= 0 {
 		e.Alive = false
 	}
@@ -123,7 +129,7 @@ func (e *Enemy) Update(p *Player, blocks []*Block) {
 	}
 }
 
-func (e *Enemy) TakeDamage(damage int) bool {
+func (e *Enemy) TakeDamage(damage float64) bool {
 	if !e.Alive {
 		return false
 	}
